requirements.txt
flask==3.0.3
slack-bolt==1.18.1
slack-sdk==3.27.1
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.31.0
